Crate simple_tables

The simple tables crate allows you to easily create table structures.

Examples

TODO: when sort is implemented, remove the “ignore”

#[table_row]
struct MyTableRow {
    id: u32,
    name: String,
    email: String,
    age: u32
}

impl PartialEq<Self> for MyTableRow {
    fn eq(&self, other: &Self) -> bool {
        self.id == other.id
            && self.name == other.name
            && self.email == other.email
            && self.age == other.age
    }
 }

#[table(rows = MyTableRow, uid = "id")]
struct MyTable {}//!//!

let mut table: dyn Table<MyTableRow> = MyTable::new();
table.push(MyTableRow { id: 1, name: String::from("Jimmy Page"), email: String::from("jimmy@led_zeppelin.com"), age: 77 });
table.insert_top(MyTableRow { id: 2, name: String::from("David Gilmour"), email: String::from("david.gilmour@pinkfloyd.com"), age: 75 });
table.insert(1, MyTableRow { id: 3, name: String::from("Kurt Cobain"), email: String::from("kurt@nirvana.com"), age: 27 });
table.sort();

let rows: Vec<MyTableRow> = vec![
    MyTableRow { id: 1, name: String::from("Jimmy Page"), email: String::from("jimmy@led_zeppelin.com"), age: 77 },
    MyTableRow { id: 2, name: String::from("David Gilmour"), email: String::from("david.gilmour@pinkfloyd.com"), age: 75 },
    MyTableRow { id: 3, name: String::from("Kurt Cobain"), email: String::from("kurt@nirvana.com"), age: 27 }
];
let table2 = MyTable::from_vec(&rows);

assert_eq!(table.get_rows(), table2.get_rows())

Re-exports

pub extern crate simple_tables_core as core;

pub extern crate simple_tables_derive as derive;

Modules

macros

Traits

IdTable

Defines a table with a unique identifier. This class should be implemented alongside the Table trait.

Table

A table should conform to this trait. Row is the table’s row type.

TableRow